requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2  ### Load or Read PDF
chromadb  ### Vector Embeddings
langchain
langchain_google_genai
faiss-cpu